clean up: move editor and renderer into seperate files
