+++ packages/plugin-text/src/plugins/katex/index.tsx
@@ -5,14 +5,13 @@ import {
   NodeControlsProps
 } from '../..'
 import * as React from 'react'
-import { Block, Editor, Inline, BlockJSON, InlineJSON } from 'slate'
+import { Block, Editor, Inline } from 'slate'
 import { isHotkey } from 'is-hotkey'
 
-import { Math } from './math.component'
-import { HoveringOverlay, styled, InlineCheckbox } from '@edtr-io/editor-ui'
-
-// @ts-ignore
-import MathQuill, { addStyles as addMathquillStyles } from 'react-mathquill'
+//@ts-ignore
+import { addStyles as addMathquillStyles } from 'react-mathquill'
+import { DefaultRendererComponent } from './renderer'
+import { DefaultEditorComponent } from './editor'
 
 addMathquillStyles()
 
@@ -43,234 +42,6 @@ export interface KatexPluginOptions {
   EditorComponent?: React.ComponentType<NodeEditorProps>
   RenderComponent?: React.ComponentType<NodeRendererProps>
   ControlsComponent?: React.ComponentType<NodeControlsProps>
-}
-
-const Wrapper = styled.div<{ inline: boolean }>(props => {
-  return {
-    whiteSpace: undefined,
-    overflowWrap: undefined,
-    ...(props.inline
-      ? {
-          display: 'inline-block'
-        }
-      : {
-          display: 'flex',
-          flexDirection: 'column',
-          alignItems: 'center',
-          marginTop: '15px'
-        })
-  }
-})
-
-const DefaultEditorComponent: React.FunctionComponent<
-  NodeEditorProps
-> = props => {
-  const { attributes, node, editor, readOnly } = props
-
-  const { data } = node as Block | Inline
-  const inline = data.get('inline')
-  const formula = data.get('formula')
-
-  function setFormula(value: string) {
-    editor.setNodeByKey(node.key, {
-      type: node.type,
-      data: {
-        formula: value,
-        inline: node.data.get('inline')
-      }
-    })
-  }
-
-  const [useVisual, setUseVisual] = React.useState(true)
-
-  const mathQuillRef = React.createRef<typeof MathQuill>()
-  const latexInputRef = React.useRef<
-    HTMLInputElement | HTMLTextAreaElement | null
-  >()
-
-  const wrappedMathquillRef = Object.defineProperty({}, 'current', {
-    // wrapper around Mathquill component
-    get: () => {
-      return mathQuillRef.current ? mathQuillRef.current.element : null
-    }
-  })
-  const [formulaState, setFormulaState] = React.useState(formula)
-
-  // if math formula gets selected or leaves edit, update formula from state
-  const edit = React.useRef(
-    props.isSelected && editor.value.selection.isCollapsed && !readOnly
-  )
-  if (
-    edit.current !==
-    (props.isSelected && editor.value.selection.isCollapsed && !readOnly)
-  ) {
-    if (formula !== formulaState) {
-      setFormula(formulaState)
-    }
-    edit.current =
-      props.isSelected && editor.value.selection.isCollapsed && !readOnly
-  }
-
-  function checkLeaveLatexInput(e: React.KeyboardEvent) {
-    if (!latexInputRef.current) return
-    const { selectionEnd, value } = latexInputRef.current
-    if (e.key === 'ArrowLeft' && selectionEnd === 0) {
-      // leave left
-      editor
-        .moveToStart()
-        .moveBackward(1)
-        .focus()
-    } else if (e.key === 'ArrowRight' && selectionEnd === value.length) {
-      // leave right
-      editor
-        .moveToEnd()
-        .moveForward(1)
-        .focus()
-    }
-  }
-
-  if (edit.current) {
-    return (
-      <Wrapper
-        {...attributes}
-        onClick={e => {
-          e.stopPropagation()
-        }}
-        inline={inline}
-      >
-        {useVisual ? (
-          <MathQuill
-            latex={formulaState} // Initial latex value for the input field
-            onChange={(latex: string) => {
-              // Called everytime the input changes
-              setFormulaState(latex)
-            }}
-            config={{
-              handlers: {
-                moveOutOf: (dir: number) => {
-                  if (dir == 1) {
-                    // leave right
-                    editor
-                      .moveToEnd()
-                      .moveForward(1)
-                      .focus()
-                  } else if (dir == -1) {
-                    // leave left
-                    editor
-                      .moveToStart()
-                      .moveBackward(1)
-                      .focus()
-                  }
-                }
-              }
-            }}
-            ref={mathQuillRef}
-            mathquillDidMount={(x: {
-              latex: () => string
-              focus: () => void
-            }) => {
-              if (x) {
-                if (x.latex() == '' && formula != '') {
-                  // Error occured
-                  alert('Error while parsing LaTeX.')
-                  setUseVisual(false)
-                }
-                setTimeout(() => {
-                  editor.blur()
-                  setTimeout(() => {
-                    x.focus()
-                  })
-                })
-              }
-            }}
-          />
-        ) : inline ? (
-          <input
-            ref={ref => (latexInputRef.current = ref)}
-            type="text"
-            value={formulaState}
-            onChange={e => {
-              //cant set formula directly, because otherwise focus jumps to end of input field
-              setFormulaState(e.target.value)
-            }}
-            onKeyDown={checkLeaveLatexInput}
-            autoFocus
-          />
-        ) : (
-          <textarea
-            style={{ width: '100%' }}
-            ref={ref => (latexInputRef.current = ref)}
-            value={formulaState}
-            onChange={e => setFormulaState(e.target.value)}
-            onKeyDown={checkLeaveLatexInput}
-            autoFocus
-          />
-        )}
-        <HoveringOverlay
-          position={'above'}
-          anchor={useVisual ? wrappedMathquillRef : latexInputRef}
-        >
-          <select
-            value={useVisual ? 'visual' : 'latex'}
-            style={{ color: 'black' }}
-            onChange={e => {
-              setUseVisual(e.target.value == 'visual')
-            }}
-          >
-            <option value="visual">visual</option>
-            <option value="latex">latex</option>
-          </select>
-          <InlineCheckbox
-            label="Inline"
-            checked={inline}
-            onChange={checked => {
-              const newData = { formula, inline: checked }
-
-              editor.removeNodeByKey(node.key)
-              if (checked) {
-                editor.insertInline({
-                  type: katexInlineNode,
-                  data: newData
-                })
-              } else {
-                editor.insertBlock({
-                  type: katexBlockNode,
-                  data: newData
-                })
-              }
-            }}
-          />
-        </HoveringOverlay>
-      </Wrapper>
-    )
-  }
-
-  return (
-    <span {...attributes}>
-      {formula ? (
-        <Math formula={formula} inline={inline} />
-      ) : (
-        <span style={{ backgroundColor: 'lightgrey' }}>[neue Formel]</span>
-      )}
-    </span>
-  )
-}
-
-class DefaultRendererComponent extends React.Component<NodeRendererProps> {
-  public render() {
-    const { node } = this.props
-
-    const { data } = node as BlockJSON | InlineJSON
-
-    if (!data) {
-      return null
-    }
-
-    const formula = data.formula as string
-    const inline = data.inline as boolean
-
-    return <Math formula={formula} inline={inline} />
-  }
 }
 
 export const createKatexPlugin = ({
